+++ docs/source/CHANGELOG.rst
@@ -4,4 +4,12 @@ github.
 Unrealeased
 ~~~~~~~~~~~
 
-- Bug Fix in param_plotter()
+Version 1.1.0
+~~~~~~~~~~~~~
+
+- Two bug fixes in param_plotter()
+- Extension of param_plotter() function to plot data, fit and residuals
+  alongside the parameter space if required.
+- Extension of param_plotter() to allow for highlighting of central
+  regions in each panel if required.
+- Inclusion of some theory into the documentation